Extract framework specific functionality into framework module & class
