+++ connexion/apis/flask_api.py
@@ -8,20 +8,21 @@ import typing as t
 import flask
 from flask import Response as FlaskResponse
 
-from connexion.apis import flask_utils
 from connexion.apis.abstract import AbstractAPI
 from connexion.decorators import SyncDecorator
-from connexion.http_facts import FORM_CONTENT_TYPES
+from connexion.frameworks import flask as flask_utils
+from connexion.frameworks.flask import Flask as FlaskFramework
 from connexion.jsonifier import Jsonifier
-from connexion.lifecycle import ConnexionRequest
 from connexion.operations import AbstractOperation
 from connexion.uri_parsing import AbstractURIParser
-from connexion.utils import is_json_mimetype
 
 logger = logging.getLogger("connexion.apis.flask_api")
 
 
 class FlaskApi(AbstractAPI):
+
+    jsonifier = Jsonifier(flask.json, indent=2)
+
     def _set_base_path(self, base_path):
         super()._set_base_path(base_path)
         self._set_blueprint()
@@ -56,63 +57,6 @@ class FlaskApi(AbstractAPI):
         self.blueprint.add_url_rule(
             flask_path, endpoint_name, endpoint, methods=[method]
         )
-
-    @classmethod
-    def is_framework_response(cls, response):
-        """Return True if provided response is a framework type"""
-        return flask_utils.is_flask_response(response)
-
-    @classmethod
-    def connexion_to_framework_response(cls, response):
-        """Cast ConnexionResponse to framework response class"""
-        return cls.build_response(
-            content_type=response.content_type,
-            headers=response.headers,
-            status_code=response.status_code,
-            data=response.body,
-        )
-
-    @classmethod
-    def build_response(
-        cls,
-        data,
-        content_type=None,
-        headers=None,
-        status_code=None,
-    ):
-        if cls.is_framework_response(data):
-            return flask.current_app.make_response((data, status_code, headers))
-
-        kwargs = {
-            "mimetype": content_type,
-            "headers": headers,
-            "response": data,
-            "status": status_code,
-        }
-        kwargs = {k: v for k, v in kwargs.items() if v is not None}
-        return flask.current_app.response_class(**kwargs)
-
-    @staticmethod
-    def get_request(*, uri_parser: AbstractURIParser, **kwargs) -> ConnexionRequest:  # type: ignore
-        return ConnexionRequest(flask.request, uri_parser=uri_parser)
-
-    @staticmethod
-    def get_body(request: ConnexionRequest) -> t.Any:
-        """Get body from a sync request based on the content type."""
-        if is_json_mimetype(request.content_type):
-            return request.get_json(silent=True)
-        elif request.mimetype in FORM_CONTENT_TYPES:
-            return request.form
-        else:
-            # Return explicit None instead of empty bytestring so it is handled as null downstream
-            return request.get_data() or None
-
-    @classmethod
-    def _set_jsonifier(cls):
-        """
-        Use Flask specific JSON loader
-        """
-        cls.jsonifier = Jsonifier(flask.json, indent=2)
 
 
 class FlaskOperation:
@@ -150,10 +94,11 @@ class FlaskOperation:
         decorator = SyncDecorator(
             self._operation,
             uri_parser_cls=self.uri_parser_class,
-            framework=self.api,
+            framework=FlaskFramework,
             parameter=True,
             response=True,
             pythonic_params=self.pythonic_params,
+            jsonifier=self.api.jsonifier,
         )
         return decorator(self._fn)
