+++ connexion/middleware/routing.py
@@ -1,6 +1,4 @@
-import functools
 import pathlib
-import re
 import typing as t
 from contextvars import ContextVar
 
@@ -9,6 +7,7 @@ from starlette.routing import Router
 from starlette.types import ASGIApp, Receive, Scope, Send
 
 from connexion.apis import AbstractRoutingAPI
+from connexion.frameworks import starlette as starlette_utils
 from connexion.middleware.abstract import ROUTING_CONTEXT, AppMiddleware
 from connexion.operations import AbstractOperation
 from connexion.resolver import Resolver
@@ -80,7 +79,7 @@ class RoutingAPI(AbstractRoutingAPI):
             operation, next_app=self.next_app
         )
         types = operation.get_path_parameter_types()
-        path = starlettify_path(path, types)
+        path = starlette_utils.starlettify_path(path, types)
         self._add_operation_internal(method, path, routing_operation)
 
     def _add_operation_internal(
@@ -99,8 +98,12 @@ class RoutingMiddleware(AppMiddleware):
         self.app = app
         # Pass unknown routes to next app
         self.router = Router(default=RoutingOperation(None, self.app))
-        starlette.convertors.register_url_convertor("float", FloatConverter())
-        starlette.convertors.register_url_convertor("int", IntegerConverter())
+        starlette.convertors.register_url_convertor(
+            "float", starlette_utils.FloatConverter()
+        )
+        starlette.convertors.register_url_convertor(
+            "int", starlette_utils.IntegerConverter()
+        )
 
     def add_api(
         self,
@@ -136,46 +139,3 @@ class RoutingMiddleware(AppMiddleware):
         # Needs to be set so starlette router throws exceptions instead of returning error responses
         scope["app"] = self
         await self.router(scope, receive, send)
-
-
-PATH_PARAMETER = re.compile(r"\{([^}]*)\}")
-PATH_PARAMETER_CONVERTERS = {"integer": "int", "number": "float", "path": "path"}
-
-
-def convert_path_parameter(match, types):
-    name = match.group(1)
-    swagger_type = types.get(name)
-    converter = PATH_PARAMETER_CONVERTERS.get(swagger_type)
-    return f'{{{name.replace("-", "_")}{":" if converter else ""}{converter or ""}}}'
-
-
-def starlettify_path(swagger_path, types=None):
-    """
-    Convert swagger path templates to flask path templates
-
-    :type swagger_path: str
-    :type types: dict
-    :rtype: str
-
-    >>> starlettify_path('/foo-bar/{my-param}')
-    '/foo-bar/{my_param}'
-
-    >>> starlettify_path('/foo/{someint}', {'someint': 'int'})
-    '/foo/{someint:int}'
-    """
-    if types is None:
-        types = {}
-    convert_match = functools.partial(convert_path_parameter, types=types)
-    return PATH_PARAMETER.sub(convert_match, swagger_path)
-
-
-class FloatConverter(starlette.convertors.FloatConvertor):
-    """Starlette converter for OpenAPI number type"""
-
-    regex = r"[+-]?[0-9]*(\.[0-9]*)?"
-
-
-class IntegerConverter(starlette.convertors.IntegerConvertor):
-    """Starlette converter for OpenAPI integer type"""
-
-    regex = r"[+-]?[0-9]+"
